+++ demo/auth-starter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # mandu-auth-starter
 
+## 0.1.68
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mandujs/core@0.54.17
+
 ## 0.1.67
 
 ### Patch Changes